requirements.txt
transformers
safetensors
opencv-python
Pillow
